requirements.txt
Flask>=1.1.1
requests>=2.18
flake8>=3.7.8
invoke>=1.3.0
pycrypto>=2.6
tinydb>=3.13.0


